requirements.txt
absl-py==0.7.1
astor==0.7.1
chardet==3.0.4
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
gast==0.2.2
glob3==0.0.1
grpcio==1.16.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.1.0
mock==2.0.0
moviepy==1.0.0
networkx==2.3
numpy==1.16.3
olefile==0.46
opencv-contrib-python==4.1.0.25
pandas==0.24.2
pbr==5.1.3
Pillow==6.0.0
PIMS==0.4.1
pprint==0.1
proglog==0.1.9
protobuf==3.7.1
pyparsing==2.4.0
pyreadline==2.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
requests==2.22.0
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
slicerator==1.0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.2
tqdm==4.32.1
urllib3==1.25.3
Werkzeug==0.15.2
wincertstore==0.2
